backend/requirements.txt
fastapi
uvicorn
opencv-python-headless
Pillow
numpy
tensorflow
roboflow
supervision
pandas
jinja2
python-multipart
starlette
python